feat: Refactor DB communication, remove moka cache (#51)
To accommodate the stringent requirement for less memory usage, the inefficient cache is no longer used.
Now random generation is done directly through queries on a "dynamic" table created at startup of the application, aggregating data from all the other tables (kinda violates the normal form, but we need it for efficient I/O and that's why the table is dynamic and not static).

* change optional data, now listing return always the least possible (core+variant), while for a single get it will be possible to choose what to get with granularity
* Move to i64/f64 for every number type, wasting some memory but removing a lot of critical conversions between types
* Create a build phase, used to initialize a SQL table 'CREATURE_CORE' using data found in other DB's tables
* Remove Creature 'from' bloated code. Now all the Creature structs must be initialized directly
* remove creature info, squished the data in the core (roles)
* remove create and scales cache. cache is now for the list of runtime values and the result of the get_all_creatures_core_data function
* remove rawcreature, a long time coming.
* add raw_query_builder.rs, helps building raw statement (no compile time checks) in a standardized way. This is done to quarantine the queries that are not statically checked
* update dependencies, remove bad implementation of validator length
* move cache to proxy.rs
* better naming convention. methods that query the DB are now "fetch" while methods that layer on top of them are "get" because the end user does not care where the data comes from.
* downgrade msrv
* limit db fetch for random generator to 20 elements
* fetch random elements for random generator from db (so it does not always return the same 20 elements)

closes: #50

+++ src/models/items/spell_caster_entry.rs
@@ -6,7 +6,7 @@ pub struct SpellCasterEntry {
     pub spell_casting_name: Option<String>,
     pub is_spell_casting_flexible: Option<bool>,
     pub type_of_spell_caster: Option<String>,
-    pub spell_casting_dc_mod: Option<i8>,
-    pub spell_casting_atk_mod: Option<i8>,
+    pub spell_casting_dc_mod: Option<i64>,
+    pub spell_casting_atk_mod: Option<i64>,
     pub spell_casting_tradition: Option<String>,
 }
