+++ src/models/scales_struct/creature_scales.rs
@@ -14,16 +14,16 @@ use std::collections::HashMap;
 
 #[derive(Default, Eq, PartialEq, Clone)]
 pub struct CreatureScales {
-    pub ability_scales: HashMap<i8, AbilityScales>,
-    pub ac_scales: HashMap<i8, AcScales>,
-    pub area_dmg_scales: HashMap<i8, AreaDmgScales>,
-    pub hp_scales: HashMap<i8, HpScales>,
+    pub ability_scales: HashMap<i64, AbilityScales>,
+    pub ac_scales: HashMap<i64, AcScales>,
+    pub area_dmg_scales: HashMap<i64, AreaDmgScales>,
+    pub hp_scales: HashMap<i64, HpScales>,
     pub item_scales: HashMap<String, ItemScales>,
-    pub perception_scales: HashMap<i8, PerceptionScales>,
-    pub res_weak_scales: HashMap<i8, ResWeakScales>,
-    pub saving_throw_scales: HashMap<i8, SavingThrowScales>,
-    pub skill_scales: HashMap<i8, SkillScales>,
-    pub spell_dc_and_atk_scales: HashMap<i8, SpellDcAndAtkScales>,
-    pub strike_bonus_scales: HashMap<i8, StrikeBonusScales>,
-    pub strike_dmg_scales: HashMap<i8, StrikeDmgScales>,
+    pub perception_scales: HashMap<i64, PerceptionScales>,
+    pub res_weak_scales: HashMap<i64, ResWeakScales>,
+    pub saving_throw_scales: HashMap<i64, SavingThrowScales>,
+    pub skill_scales: HashMap<i64, SkillScales>,
+    pub spell_dc_and_atk_scales: HashMap<i64, SpellDcAndAtkScales>,
+    pub strike_bonus_scales: HashMap<i64, StrikeBonusScales>,
+    pub strike_dmg_scales: HashMap<i64, StrikeDmgScales>,
 }
